Major update to the file structure and server api endpoints
